pyproject.toml
[project]
name = "spikeai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.125.0",
    "google-analytics-data>=0.19.0",
    "langchain-openai>=1.1.5",
    "langgraph>=1.0.5",
    "openai>=2.13.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
